requirements.txt
aiohttp==3.8.4
anticaptchaofficial==1.0.44
beautifulsoup4==4.10.0
cryptography==3.4.7
Fernet==1.0.1
requests==2.26.0
retry==0.9.2
selenium==4.1.0
tenacity==8.0.1
undetected-chromedriver==3.1.5.post4
webdriver-manager==3.5.2
aiofiles 
cloudscraper
aiolimiter
psutil
elasticsearch
sentry_sdk
prometheus_client
fake_useragent
browsermob_proxy
Pillow
numpy
scipy
opencv-python
asyncio
chromedriver-autoinstaller
python-dotenv